Log in user
A user can sign up and there information will be saved in the database and redirected to the log in page.
We need to figure out what we want to be saved in the user login.

Still need to figure out the specific id to use for each user.

A user should be able to log in and there information will be retrieved from the server.

They're information will also be saved into local stored so they stay logged into the page if they leave.
The user can sign out and that will remove the save values in local storage.

There is a new page to load the database that is required and used in server to  use the database. This can be changed later if need be, but was needed
for testing the log in and log out features.

Little to no styling at this time. I will be working on that today and also more testing to make sure everything with the log in and out works.